+++ ReleaseNotes.md
@@ -8,6 +8,13 @@
 - Virtues and flaws are working again on beast sheet
 - Vis source attributes are visible again
 - Quick dialogs (Vitals, Combat, Magic) are now updating when the linked character changes (V12 only)
+- [V12] Changed actor links mechanics since appId of a sheet is only defined on render (instead of constructor).
+- Covenant laborers were wrongly considered servants
+- Covenant teamsters were not considered workers.
+- It is possible again to go to previous topic of a book in scriptorium
+- Missing spell warning is now properly formated in scriptorium
+- Multitopic book reading was not working after a change to the dialog (reader, date, ...)
+- There is no longer an error when you try to read an ability or spell mastery book you are too skilled for.
 
 ## 2.3.0.10, Quendalon, raised by the Fae BF3